requirements.txt
annotated-types==0.7.0
anyio==4.4.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
et-xmlfile==1.1.0
exceptiongroup==1.2.2
fastapi==0.112.2
google-api-core==2.24.0
google-auth==2.37.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
googleapis-common-protos==1.66.0
granian==1.5.2
greenlet==3.0.3
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
idna==3.8
numpy==2.1.0
openpyxl==3.1.5
pandas==2.2.2
proto-plus==1.25.0
protobuf==5.29.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
python-http-client==3.3.7
python-multipart==0.0.9
pytz==2024.1
requests==2.32.3
rsa==4.9
sendgrid==6.11.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
starkbank-ecdsa==2.2.0
starlette==0.38.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
uvicorn==0.30.6
